refactor: apply coderabbit feedback on drawer implementation
