+++ src/features/drawer/DrawerMenu.module.css
@@ -5,6 +5,13 @@
   background: rgba(36, 49, 48, 0.45);
   display: flex;
   align-items: flex-start;
+  visibility: hidden;
+  pointer-events: none;
+}
+
+.overlayVisible {
+  visibility: visible;
+  pointer-events: auto;
 }
 
 .drawer {
